RingCT inputs organized with related data
Visually, it didn't look great at all to lump all the commitment, inputs, and key_images together when they are related to one another in groups. This is meant to visually represent that accurately.  This should make it easier to turn the Json into objects from other clients and make representation in explorers better.