[Chain] Prune entire orphaned chains.
If the index contains forks other than the active fork, then
removing one from an early part of the chain can cause startup to
assert `pindexWalk->pprev` on the remaining blocks in the fork which
no longer have parents in the index. (Issue #692)

This is made likely by the pruning process that only prunes orphans
based on height far enough from the best block.

Solution: After finding out which index entries are old enough to be
pruned, search the index for all entries that descend from any of those,
and add them to be pruned.

Additionally, CBlockIndex pointers are passed all over the codebase and
are never deleted, but their hash values are pointers to the
mapBlockIndex's keys. Normally this isn't a problem (unordered_map),
until the data is erased and the memory reused.