[Chain] Keep hashAccumulators more up-to-date.
Use hashAccumulators instead of re-hashing mapAccumulatorHashes, as sometimes the map
is not used but the cached hash value it had is (e.g. constructing the
CBlockIndex from a CBlockHeader).

Update hashAccumulators on any calls to AddAccumulators.
Copy it when copying mapBlockIndex.
Finally, copy it on loading from the diskindex--I believe this is
the cause of some indexes (particularly orphans) having inconsistent hashes.

Fixes #942.