Create multiple transactions with one command.
RPC send commands have a new argument now that controls whether
the amount specified can be split across multiple transactions.
It defaults to single-transaction mode for now.

SelectCoinsForOneTx: use a sliding window algorithm to pick inputs.
This will pick one small input if possible, so that if change is
produced it does not result in an ever-growing pile of tiny utxos.