Update README.md
Updated the description of Veil from what was the perspective just before launch, to what Veil is nearly three years later.
We're not just "intending to" set a standard, it has been done. The zerocoin and mining protocols have been changed since then, and the budget structure has been changed.