Proposed fix for force_return errors.
We do not need to look up an address in the address book map
just to check the type of the key in the map, as the type
must be exactly the same as the provided destination. But if we don't
find the destination in the map, we get back `end()` which is not
a valid value and may have a `which_` field in the `boost::variant`
that can trigger the assert.