Fix displayed tx value when wallet gets unlocked while tx unconfirmed
AnonWallet::RescanWallet did not unblind CT / RingCT outputs of
unconfirmed transactions. Thus, when such a transaction is received it
is shown with "Hidden" value, but if the user unlocked the wallet while
the transaction is still unconfirmed, the value (in the GUI) wasn't
updated even when the transaction gets confirmed. Only after a wallet
restart the correct value was displayed.

With this commit, the correct value is displayed directly after
unlocking the wallet, also if the transaction is still unconfirmed.