Fix possible deadlock situation when encrypting the wallet
CWallet::EncryptWallet doesn't release its `cs_wallet` lock before
calling CWallet::Unlock, which ends up calling AnonWallet::UnlockWallet,
which does the partial wallet rescan and for that acquires locks
`cs_main` and `cs_wallet`. Thus, the order of lock acquisition in this
case is `cs_wallet`, `cs_main`. However, across the entire application
consistently the order `cs_main`, `cs_wallet` is used. This leads to a
deadlock if another thread already holds `cs_main` and is waiting for
`cs_wallet`, but the AnonWallet::UnlockWallet currently has `cs_wallet`,
and is waiting for `cs_main`.

Since rescaning the anon wallet is anyway not required when encrypting
the wallet (because it restarts the wallet and then the user has to
unlock it again), the simple solution was to not do this rescan when
unlocking during wallet encryption, thus there is no need to acquire
`cs_main`.