Fix wrong CT / RingCT amounts in transaction history when wallet locked (#1001)
* Fix wrong CT amounts in transaction history when wallet locked

- When a CT transaction arrives while the wallet is locked (or unlocked
  for staking only):
  - Instead of a wrong amount show "Hidden" in the transaction history
    entry.
  - In the tooltip of the transaction row show the message "Unlock wallet
    to see hidden amount".
- When the wallet is then unlocked, automatically do a
  `rescanringctwallet`.
- Fix `rescanringctwallet` to properly update corresponding entries in
  the transaction history GUI.

* Mark AnonWallet::RescanWallet with EXCLUSIVE_LOCKS_REQUIRED annotations

* Fix not showing "Hidden" on incoming zerocoin spend CT transactions

If the incoming CTs come from a zerocoin spend and the transaction
arrived while the wallet was locked, they were shown in the transaction
history with an amount of "0.00000000" instead of "Hidden".

* Fix missing `ORF_OWNED` flag for incoming RingCT while wallet locked

This caused such transactions to show up in the GUI as "(n/a)" if the
wallet was locked when the transaction arrived.

* Also show "Hidden" in the GUI notification for locked incoming transactions

Show "Hidden" instead of wrong amounts like "0.00000000" in the "Incoming
transaction" GUI notification that is shown to the user.

* Fix displayed tx value when wallet gets unlocked while tx unconfirmed

AnonWallet::RescanWallet did not unblind CT / RingCT outputs of
unconfirmed transactions. Thus, when such a transaction is received it
is shown with "Hidden" value, but if the user unlocked the wallet while
the transaction is still unconfirmed, the value (in the GUI) wasn't
updated even when the transaction gets confirmed. Only after a wallet
restart the correct value was displayed.

With this commit, the correct value is displayed directly after
unlocking the wallet, also if the transaction is still unconfirmed.

* Optimization: keep track of locked transactions

Keep track of locked CT/RingCT transactions that arrive while the wallet
is locked and contain blinded values, in order to only rescan those
transactions when unlocking the wallet instead of doing a full rescan
each time.

* Fix possible deadlock situation when encrypting the wallet

CWallet::EncryptWallet doesn't release its `cs_wallet` lock before
calling CWallet::Unlock, which ends up calling AnonWallet::UnlockWallet,
which does the partial wallet rescan and for that acquires locks
`cs_main` and `cs_wallet`. Thus, the order of lock acquisition in this
case is `cs_wallet`, `cs_main`. However, across the entire application
consistently the order `cs_main`, `cs_wallet` is used. This leads to a
deadlock if another thread already holds `cs_main` and is waiting for
`cs_wallet`, but the AnonWallet::UnlockWallet currently has `cs_wallet`,
and is waiting for `cs_main`.

Since rescaning the anon wallet is anyway not required when encrypting
the wallet (because it restarts the wallet and then the user has to
unlock it again), the simple solution was to not do this rescan when
unlocking during wallet encryption, thus there is no need to acquire
`cs_main`.

Co-authored-by: Johannes Westphal <jojo@w-hat.de>