Commit for release 1.4.0.0
This release is MANDATORY due to the consensus change in ProgPoW (Proof of Work) mining reduction of the DAG size to allow GPU miners with lower RAM sizes to mine.

This will take place at block 2,100,000, expected the 21st of January, or possibly on the 20th.

This wallet also serves as an update directly instead of v1.3.0.0 from the 1.2.x wallet, as it contains the same code that is in mandatory version 1.3.0.0, meaning that you can go directly from version 1.2 to 1.4. 1.3 and this 1.4 extends Proof of Work mining preventing it from ending at block two million expected on the 12th of November 2022. Proof of Work mining will now continue throughout Veil's future until the block subsidy and the Supply of Veil reaches its maximum, expected in January 2038, after which Veil nodes will be incentivised by transaction fees.